+++ src/main/java/dev/dsf/bpe/util/task/input/generator/DownloadedBytesGenerator.java
@@ -1,6 +1,6 @@
 package dev.dsf.bpe.util.task.input.generator;
 
-import org.hl7.fhir.r4.model.IntegerType;
+import org.hl7.fhir.r4.model.DecimalType;
 import org.hl7.fhir.r4.model.Task;
 
 import dev.dsf.bpe.ConstantsPing;
@@ -11,10 +11,10 @@ public final class DownloadedBytesGenerator
 	{
 	}
 
-	public static Task.ParameterComponent create(int bytes)
+	public static Task.ParameterComponent create(long bytes)
 	{
 		Task.ParameterComponent param = new Task.ParameterComponent();
-		param.setValue(new IntegerType(bytes)).getType().addCoding().setSystem(ConstantsPing.CODESYSTEM_DSF_PING)
+		param.setValue(new DecimalType(bytes)).getType().addCoding().setSystem(ConstantsPing.CODESYSTEM_DSF_PING)
 				.setCode(ConstantsPing.CODESYSTEM_DSF_PING_VALUE_DOWNLOADED_BYTES);
 		return param;
 	}
