+++ src/main/java/dev/dsf/bpe/util/task/input/generator/NetworkSpeedMetricGenerator.java
@@ -2,7 +2,6 @@ package dev.dsf.bpe.util.task.input.generator;
 
 import org.hl7.fhir.r4.model.Coding;
 import org.hl7.fhir.r4.model.DecimalType;
-import org.hl7.fhir.r4.model.IntegerType;
 import org.hl7.fhir.r4.model.Task;
 
 import dev.dsf.bpe.ConstantsPing;
@@ -22,10 +21,10 @@ public final class NetworkSpeedMetricGenerator
 		return downloadedDuration;
 	}
 
-	public static Task.ParameterComponent createDownloadedBytes(int bytes)
+	public static Task.ParameterComponent createDownloadedBytes(long bytes)
 	{
 		Task.ParameterComponent downloadedBytes = new Task.ParameterComponent();
-		downloadedBytes.setValue(new IntegerType(bytes)).getType().addCoding(new Coding(
+		downloadedBytes.setValue(new DecimalType(bytes)).getType().addCoding(new Coding(
 				ConstantsPing.CODESYSTEM_DSF_PING, ConstantsPing.CODESYSTEM_DSF_PING_VALUE_DOWNLOADED_BYTES, null));
 		return downloadedBytes;
 	}
